Persist user settings to app.config.
Things like the following are persisted
- the color of the brush used to draw/measure/write text
- the size of the brush used to draw
- the font used
- the default unit and ratio used

A bunch of cleanup work around the helper code (splitting to individual files, moving to a new folder)

Re-create and move the logo source file to a separate folder.
It was broken by various changes and needed to be re-created.

Improve the new document dialog to make it clear how the conversion rate works.

Include the Style of the font in the information that is persisted in the project file.