This change introduces a way to store images that does not duplicate the original image every time.
In order to achieve this, we introduce a new object on the Drawing type (Images).
This type will store a map between the image file and a key.

When an image needs to draw the data, it will resolve the image (via a resolver that the Drawing object provides) that will return the actual data.

This reduces the size from N * image size to image size.